% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Clean latitude, longitude, and date in NOAA significant earthquake database}
\usage{
eq_clean_data(noaa_data)
}
\arguments{
\item{noaa_data}{NOAA significant earthquakes dataset}
}
\value{
A data.frame, data.table with updated longitude, latitude, and date
 variables.
}
\description{
A function that builds a date field and ensures that the longitude and
 latitude of the earthquake location are numeric.
The function takes in the NOAA data, which is obtained from
 \url{https://www.ngdc.noaa.gov/nndc/struts/form?t=101650&s=1&d=1}) and
 returns a data frame with a valid date field and numeric longitude and
 latitude.
}
\section{Imported functions}{

 data.table syntax
}

\examples{
library(data.table)
raw_noaa <- as.data.table(noaa_data)
clean_noaa <- eq_clean_data(raw_noaa)

}
