% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Build an html label for earthquakes}
\usage{
eq_create_label(earthquake_data)
}
\arguments{
\item{earthquake_data}{A dataframe containing location names in column
\code{LOCATION_NAME}, earthquake magnitude in column \code{EQ_PRIMARY},
and total deaths caused by the earthquake in column \code{TOTAL_DEATHS}. 
An example of a correctly formatted dataset is noaa_data, which is included
in the package.}
}
\value{
A character vector of formatted html strings to label the popups in
  \code{\link{eq_map}}
}
\description{
Create a nicely formatted HTML label for use with \code{\link{eq_map}}
 showing Location, Magnatude, and Total Deaths.  Missing values are skipped
 in the created label.
}
\examples{
library(data.table)
raw_noaa <- as.data.table(noaa_data)
raw_noaa \%>\%
  eq_clean_data() \%>\%
  dplyr::filter(COUNTRY == "MEXICO" & lubridate::year(DATE) >= 2000) \%>\%
  eq_map(annot_col = "DATE")
## Plot earthquakes in Mexico with nicely formated popup windows that include location, magnitude, and total deaths for each earthquake
raw_noaa \%>\%
  eq_clean_data() \%>\%
  eq_location_clean() \%>\%
  dplyr::filter(COUNTRY == "MEXICO" & lubridate::year(DATE) >= 2000) \%>\%
  dplyr::mutate(popup_text = eq_create_label(.)) \%>\%
  eq_map(annot_col = "popup_text")

}
\seealso{
\code{\link{eq_map}}, \pkg{leaflet}
}
