% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Clean location in NOAA significant earthquake database}
\usage{
eq_location_clean(noaa_data)
}
\arguments{
\item{noaa_data}{NOAA significant Earthquakes dataset}
}
\value{
A data.frame, data.table with updated location name
}
\description{
A function that removes country information from the location field and
 ensures that the location is in title case.
The function takes in the NOAA data, which is obtained from
 \url{https://www.ngdc.noaa.gov/nndc/struts/form?t=101650&s=1&d=1}) and
 returns a data frame with a cleaned location name
}
\section{Imported functions}{

 data.table syntax
}

\examples{
library(data.table)
raw_noaa <- as.data.table(noaa_data)
clean_noaa <- eq_location_clean(raw_noaa)

}
