% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet.R
\name{eq_map}
\alias{eq_map}
\title{Create an interactive map of historical earthquakes}
\usage{
eq_map(earthquake_data, annot_col)
}
\arguments{
\item{earthquake_data}{A dataframe with earthquake location columns
\code{LATITUDE} and \code{LONGITUDE}, and magnatude column
\code{EQ_PRIMARY}. The included noaa_data dataframe is already
 correctly formateed, and is intended to be used with this function}

\item{annot_col}{The character name of the column in \code{data} containing
a character vector of optional popup text to be shown when an earthquake
is clicked on.}
}
\value{
An html map object, to which further Leaflet objects can be added
}
\description{
Create an interactive map of historical earthquakes using the NOAA
 earthquake database included with this package.  Earthquakes are plotted
 as circles with their radii proportional to the magnatude of the
 earthquake. Optionally, labels can be passed which will pop up when the
 earthquake is clicked on.
}
\details{
This function uses of the popular \code{leaflet} package, which creates
 interactive html maps within R
}
\examples{
library(data.table)
raw_noaa <- as.data.table(noaa_data)
raw_noaa \%>\%
  eq_clean_data() \%>\%
  dplyr::filter(COUNTRY == "MEXICO" & lubridate::year(DATE) >= 2000) \%>\%
  eq_map(annot_col = "DATE")
## Plot earthquakes in Mexico with nicely formated popup windows that include location, magnitude, and total deaths for each earthquake
raw_noaa \%>\%
  eq_clean_data() \%>\%
  eq_location_clean() \%>\%
  dplyr::filter(COUNTRY == "MEXICO" & lubridate::year(DATE) >= 2000) \%>\%
  dplyr::mutate(popup_text = eq_create_label(.)) \%>\%
  eq_map(annot_col = "popup_text")

}
\seealso{
\pkg{leaflet}
}
